add test for write_empty_chunks
