+++ tests/benchmarks/test_read_zarr_python_benchmark.py
@@ -1,5 +1,5 @@
 import pytest
-from tests.benchmark_parameters import (
+from tests.benchmarks.benchmark_parameters import (
     CHUNK_SIZE,
     BLOSC_CLEVEL,
     BLOSC_CNAME,
@@ -7,14 +7,13 @@ from tests.benchmark_parameters import (
     GZIP_LEVEL,
     ZSTD_LEVEL,
 )
-import read_write_tensorstore
 
 try:
     import read_write_zarr_v3 as read_write_zarr
 except ImportError:
     import read_write_zarr_v2 as read_write_zarr
 
-pytestmark = [pytest.mark.tensorstore]
+pytestmark = [pytest.mark.zarr_python]
 
 
 @pytest.mark.benchmark(group="read")
@@ -33,11 +32,11 @@ def test_read_blosc(
     blosc_shuffle,
     blosc_cname,
 ):
-    blosc_compressor = read_write_tensorstore.get_blosc_compressor(
+    blosc_compressor = read_write_zarr.get_blosc_compressor(
         blosc_cname, blosc_clevel, blosc_shuffle
     )
 
-    read_write_tensorstore.write_zarr_array(
+    read_write_zarr.write_zarr_array(
         image=image,
         store_path=store_path,
         overwrite=True,
@@ -49,7 +48,7 @@ def test_read_blosc(
     benchmark.extra_info["compression_ratio"] = compression_ratio
 
     benchmark.pedantic(
-        read_write_tensorstore.read_zarr_array,
+        read_write_zarr.read_zarr_array,
         args=(store_path,),
         rounds=rounds,
         warmup_rounds=warmup_rounds,
@@ -62,9 +61,9 @@ def test_read_blosc(
 def test_read_gzip(
     benchmark, image, rounds, warmup_rounds, store_path, chunk_size, gzip_level
 ):
-    gzip_compressor = read_write_tensorstore.get_gzip_compressor(gzip_level)
+    gzip_compressor = read_write_zarr.get_gzip_compressor(gzip_level)
 
-    read_write_tensorstore.write_zarr_array(
+    read_write_zarr.write_zarr_array(
         image=image,
         store_path=store_path,
         overwrite=True,
@@ -76,7 +75,7 @@ def test_read_gzip(
     benchmark.extra_info["compression_ratio"] = compression_ratio
 
     benchmark.pedantic(
-        read_write_tensorstore.read_zarr_array,
+        read_write_zarr.read_zarr_array,
         args=(store_path,),
         rounds=rounds,
         warmup_rounds=warmup_rounds,
@@ -89,9 +88,9 @@ def test_read_gzip(
 def test_read_zstd(
     benchmark, image, rounds, warmup_rounds, store_path, chunk_size, zstd_level
 ):
-    zstd_compressor = read_write_tensorstore.get_zstd_compressor(zstd_level)
+    zstd_compressor = read_write_zarr.get_zstd_compressor(zstd_level)
 
-    read_write_tensorstore.write_zarr_array(
+    read_write_zarr.write_zarr_array(
         image=image,
         store_path=store_path,
         overwrite=True,
@@ -103,7 +102,7 @@ def test_read_zstd(
     benchmark.extra_info["compression_ratio"] = compression_ratio
 
     benchmark.pedantic(
-        read_write_tensorstore.read_zarr_array,
+        read_write_zarr.read_zarr_array,
         args=(store_path,),
         rounds=rounds,
         warmup_rounds=warmup_rounds,
