+++ tests/benchmarks/test_write_zarr_python_benchmark.py
@@ -1,9 +1,10 @@
 import pytest
-from tests.benchmark_parameters import (
+from utils import remove_output_dir
+from tests.benchmarks.benchmark_parameters import (
     CHUNK_SIZE,
     BLOSC_CLEVEL,
-    BLOSC_CNAME,
     BLOSC_SHUFFLE,
+    BLOSC_CNAME,
     GZIP_LEVEL,
     ZSTD_LEVEL,
 )
@@ -16,12 +17,12 @@ except ImportError:
 pytestmark = [pytest.mark.zarr_python]
 
 
-@pytest.mark.benchmark(group="read")
+@pytest.mark.benchmark(group="write")
 @pytest.mark.parametrize("chunk_size", CHUNK_SIZE)
 @pytest.mark.parametrize("blosc_clevel", BLOSC_CLEVEL)
 @pytest.mark.parametrize("blosc_shuffle", BLOSC_SHUFFLE)
 @pytest.mark.parametrize("blosc_cname", BLOSC_CNAME)
-def test_read_blosc(
+def test_write_blosc(
     benchmark,
     image,
     rounds,
@@ -36,74 +37,71 @@ def test_read_blosc(
         blosc_cname, blosc_clevel, blosc_shuffle
     )
 
-    read_write_zarr.write_zarr_array(
-        image=image,
-        store_path=store_path,
-        overwrite=True,
-        chunks=(chunk_size, chunk_size, chunk_size),
-        compressor=blosc_compressor,
-    )
-
-    compression_ratio = read_write_zarr.get_compression_ratio(store_path)
-    benchmark.extra_info["compression_ratio"] = compression_ratio
+    def setup():
+        remove_output_dir(store_path)
+        return (), {
+            "image": image,
+            "store_path": store_path,
+            "overwrite": False,
+            "chunks": (chunk_size, chunk_size, chunk_size),
+            "compressor": blosc_compressor,
+        }
 
     benchmark.pedantic(
-        read_write_zarr.read_zarr_array,
-        args=(store_path,),
+        read_write_zarr.write_zarr_array,
+        setup=setup,
         rounds=rounds,
         warmup_rounds=warmup_rounds,
     )
 
 
-@pytest.mark.benchmark(group="read")
+@pytest.mark.benchmark(group="write")
 @pytest.mark.parametrize("chunk_size", CHUNK_SIZE)
 @pytest.mark.parametrize("gzip_level", GZIP_LEVEL)
-def test_read_gzip(
+def test_write_gzip(
     benchmark, image, rounds, warmup_rounds, store_path, chunk_size, gzip_level
 ):
     gzip_compressor = read_write_zarr.get_gzip_compressor(gzip_level)
 
-    read_write_zarr.write_zarr_array(
-        image=image,
-        store_path=store_path,
-        overwrite=True,
-        chunks=(chunk_size, chunk_size, chunk_size),
-        compressor=gzip_compressor,
-    )
-
-    compression_ratio = read_write_zarr.get_compression_ratio(store_path)
-    benchmark.extra_info["compression_ratio"] = compression_ratio
+    def setup():
+        remove_output_dir(store_path)
+        return (), {
+            "image": image,
+            "store_path": store_path,
+            "overwrite": False,
+            "chunks": (chunk_size, chunk_size, chunk_size),
+            "compressor": gzip_compressor,
+        }
 
     benchmark.pedantic(
-        read_write_zarr.read_zarr_array,
-        args=(store_path,),
+        read_write_zarr.write_zarr_array,
+        setup=setup,
         rounds=rounds,
         warmup_rounds=warmup_rounds,
     )
 
 
-@pytest.mark.benchmark(group="read")
+@pytest.mark.benchmark(group="write")
 @pytest.mark.parametrize("chunk_size", CHUNK_SIZE)
 @pytest.mark.parametrize("zstd_level", ZSTD_LEVEL)
-def test_read_zstd(
+def test_write_zstd(
     benchmark, image, rounds, warmup_rounds, store_path, chunk_size, zstd_level
 ):
     zstd_compressor = read_write_zarr.get_zstd_compressor(zstd_level)
 
-    read_write_zarr.write_zarr_array(
-        image=image,
-        store_path=store_path,
-        overwrite=True,
-        chunks=(chunk_size, chunk_size, chunk_size),
-        compressor=zstd_compressor,
-    )
-
-    compression_ratio = read_write_zarr.get_compression_ratio(store_path)
-    benchmark.extra_info["compression_ratio"] = compression_ratio
+    def setup():
+        remove_output_dir(store_path)
+        return (), {
+            "image": image,
+            "store_path": store_path,
+            "overwrite": False,
+            "chunks": (chunk_size, chunk_size, chunk_size),
+            "compressor": zstd_compressor,
+        }
 
     benchmark.pedantic(
-        read_write_zarr.read_zarr_array,
-        args=(store_path,),
+        read_write_zarr.write_zarr_array,
+        setup=setup,
         rounds=rounds,
         warmup_rounds=warmup_rounds,
     )
